--- components/common/CharmEditor/components/handles/index.ts
@@ -1,1 +1,0 @@
-export * from './handles';
